Extract agent-task gate reporting module
